+++ src/components/Login.svelte
@@ -9,6 +9,7 @@
 
   const defaultHost = 'github.com';
   let errors: { [inputName: string]: ValidatorResult } = {};
+  let loading = false;
 
   let form: {
     [inputName: string]: {
@@ -57,7 +58,8 @@
     validateForm(data);
 
     if (isFormValid()) {
-      $auth.signIn(data);
+      loading = true;
+      $auth.signIn(data).finally(() => (loading = false));
     } else {
       console.log('Invalid Form');
     }
@@ -139,11 +141,36 @@
       </span>
     </div>
     <button
-      class="w-full text-white font-bold bg-blue-600 hover:bg-blue-800 focus:ring-4 focus:ring-blue-300 rounded-md text-sm px-4 py-2 dark:hover:bg-blue-700 focus:outline-none dark:focus:ring-blue-800"
+      class={`${
+        loading ? 'opacity-50' : ''
+      } flex justify-center items-center w-full text-white font-bold bg-blue-600 hover:bg-blue-800 focus:ring-4 focus:ring-blue-300 rounded-md text-sm px-4 py-2 dark:hover:bg-blue-700 focus:outline-none dark:focus:ring-blue-800`}
       type="submit"
       title="Submit Button"
+      disabled={loading}
     >
       Submit
+      {#if loading}
+        <svg
+          class="animate-spin h-4 w-4 text-white ml-3"
+          xmlns="http://www.w3.org/2000/svg"
+          fill="none"
+          viewBox="0 0 24 24"
+        >
+          <circle
+            class="opacity-25"
+            cx="12"
+            cy="12"
+            r="10"
+            stroke="currentColor"
+            stroke-width="4"
+          />
+          <path
+            class="opacity-75"
+            fill="currentColor"
+            d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
+          />
+        </svg>
+      {/if}
     </button>
   </form>
 </div>